--- front50-core/src/main/groovy/com/netflix/spinnaker/front50/config/PluginBinaryCacheConfiguration.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.netflix.spinnaker.front50.config;
-
-import com.netflix.spinnaker.front50.plugins.CachingPluginBinaryStorageService;
-import com.netflix.spinnaker.front50.plugins.PluginBinaryStorageService;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-@Configuration
-@ConditionalOnBean(PluginBinaryStorageService.class)
-@ConditionalOnProperty(value = "plugin-binary-cache.enabled", matchIfMissing = true)
-public class PluginBinaryCacheConfiguration {
-
-  @Primary
-  @Bean
-  PluginBinaryStorageService cachingPluginBinaryStorageService(
-      PluginBinaryStorageService pluginBinaryStorageService) {
-    return new CachingPluginBinaryStorageService(pluginBinaryStorageService);
-  }
-}
