--- front50-core/src/main/groovy/com/netflix/spinnaker/front50/plugins/PluginBinaryAlreadyExistsException.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2020 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.netflix.spinnaker.front50.plugins;
-
-import static java.lang.String.format;
-
-import com.netflix.spinnaker.kork.exceptions.UserException;
-
-/**
- * Thrown whenever an upload attempt is made on for a plugin ID and version that already has a
- * plugin binary in storage.
- */
-public class PluginBinaryAlreadyExistsException extends UserException {
-
-  public PluginBinaryAlreadyExistsException(String pluginKey) {
-    super(format("Plugin binary already exists for '%s'", pluginKey));
-  }
-}
